requirements.txt
pandas
openai
openpyxl
geopy
llama-index-llms-ollama
llama-index-core
